+++ dev/io.openliberty.mcp.internal/src/io/openliberty/mcp/internal/ToolDescription.java
@@ -47,6 +47,9 @@ public class ToolDescription {
         return annotations;
     }
 
+    /**
+     * Only for testing
+     */
     public ToolDescription(ToolMetadata toolMetadata) {
         this(toolMetadata, true);
     }
